[MULTIPLE ISSUES] Fixes for columnar datasets
- user model changes: no
- storage format changes: no
- interface changes: no
- metadata changes: yes

Details:
This patch introduces multiple changes/fixes::
- ASTERIXDB-3194: Write DatasetFormat for row datasets
  in Metadata.Dataset
- ASTERIXDB-3193: Fix bug in Pushdown rule when multi-part
  dataverse names are used
- ASTERIXDB-3192: Fix handling empty components of
  columnar datasets
- ASTERIXDB-3191: Change type of FreeSpaceTolerance to
  double (instead of float)

Metadata changes:
- DatasetFormat will appear for both row
  and columnar datasets
- FreeSpaceTolerance in Metadata.Dataset will
  be displayed differently
  - before: 0.15000000596046448
  - after: 0.15

Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/17542
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ asterixdb/asterix-column/src/main/java/org/apache/asterix/column/ColumnManagerFactory.java
@@ -45,10 +45,10 @@ public final class ColumnManagerFactory implements IColumnManagerFactory {
     private final List<Integer> keySourceIndicator;
     private final int pageSize;
     private final int maxTupleCount;
-    private final float tolerance;
+    private final double tolerance;
 
     public ColumnManagerFactory(ARecordType datasetType, ARecordType metaType, List<List<String>> primaryKeys,
-            List<Integer> keySourceIndicator, int pageSize, int maxTupleCount, float tolerance) {
+            List<Integer> keySourceIndicator, int pageSize, int maxTupleCount, double tolerance) {
         this.pageSize = pageSize;
         this.maxTupleCount = maxTupleCount;
         this.tolerance = tolerance;
